Uses mixin to reuse menu functionality
