+++ components/ui/menu/SidebarMenu.vue
@@ -41,41 +41,13 @@
 import Vue from 'vue'
 import { Component } from 'vue-property-decorator'
 import MenuLink from '~/components/ui/menu/MenuLink.vue'
-
-import {
-  ORDERED_QISKIT_ELEMENTS,
-  ORDERED_COMMUNITY_SUB_LINKS,
-  COMMUNITY_LINK,
-  TUTORIALS_LINK,
-  DOCUMENTATION_LINK,
-  NavLink
-} from '~/constants/menuLinks'
+import menuMixin from '~/mixins/menu'
 
 @Component({
-  components: { MenuLink }
+  components: { MenuLink },
+  mixins: [menuMixin]
 })
-export default class extends Vue {
-  qiskitElements: Array<NavLink> = ORDERED_QISKIT_ELEMENTS
-  learnMore: Array<NavLink> = [TUTORIALS_LINK, DOCUMENTATION_LINK]
-  communityLink: NavLink = COMMUNITY_LINK
-  communitySubLinks: Array<NavLink> = ORDERED_COMMUNITY_SUB_LINKS
-
-  isPathStartingWith (linkPath: string) {
-    return this.$route.path.startsWith(linkPath)
-  }
-
-  isActive (link: NavLink) {
-    const isCommunityLink = link.label === 'Community'
-
-    return isCommunityLink
-      ? this.communitySubLinks.some(communitySection => this.isPathStartingWith(communitySection.url))
-      : this.isPathStartingWith(link.url)
-  }
-
-  isCommunityActive () {
-    return this.communitySubLinks.some(link => this.isActive(link))
-  }
-}
+export default class extends Vue {}
 </script>
 
 <style lang="scss" scoped>
